refactor: move collection delete logic from endpoint to service layer
Extract DB queries, cancel+barrier, and async cleanup scheduling from
collections.py endpoint into CollectionService.delete() with private
helpers. Lazy-import temporal_service to avoid circular imports.
